% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CohenDLibrary.R
\docType{package}
\name{CohenDLibrary-package}
\alias{CohenDLibrary}
\alias{CohenDLibrary-package}
\title{CohenDLibrary: Cohen's D computation with confidence intervals}
\description{
Computing Cohen's d_p in between-subject and within-subject designs.
}
\details{
CohenDLibrary is a library that computes the Cohen's d along with confidence intervals. 
Its main contribution is to allow confidence intervals in the repeated-measure designs.
The main function is 

CohenD(statistics = list(...), method, design, conf.level, verbose = FALSE)  

see help(CohenD) for more

Subsidiary functions are
plambdap(x, nu, ncp) cumulative probability of the lambda prime distribution with parameters nu, ncp
dlambdap(x, nu, ncp) density
qlambdap(x, nu, ncp) quantile
 
pkprime(x, nu1, nu2, ncp) cumulative probability of the K prime distribution with parameters nu1, nu2, ncp
dkprime(x, nu1, nu2, ncp) density
qkprime(x, nu1, nu2, ncp) quantile
 
These functions are implemented from the FORTRAN source of Poitevineau & Lecoutre, 2010.
Note that the library sadists also has implementations for these two distributions. However,
the sadists::kprime distribution is innacurate for small nu1 or small nu2.
}
\keyword{internal}
