% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CohensdpInitialize.R
\name{explain}
\alias{explain}
\title{explain}
\usage{
explain(x, ...)
}
\arguments{
\item{x}{an object to explain}

\item{...}{ignored}
}
\description{
explain provides a human-readable, exhaustive, description of
the results. It also provides references to the key results.
}
