% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plambdap.R
\name{plambdap}
\alias{plambdap}
\title{lambdap}
\usage{
plambdap(x, nu, ncp)
}
\description{
plambdap(x, nu, ncp) cumulative probability of the lambda prime distribution with parameters nu, ncp
dlambdap(x, nu, ncp) density of the lambda prime distribution
qlambdap(x, nu, ncp) quantile of the lambda prime distribution
 
These functions are implemented from the FORTRAN source of Poitevineau & Lecoutre, 2010.
Note that the library sadists also has implementations for these two distributions. However,
the sadists::kprime distribution is innacurate for small nu1 or small nu2.
}
\details{
lambdap are functions that compute the Lambda prime distribution.
}
\examples{

dlambdap(6.0, 99, 6.19) # 0.436873
plambdap(6.0, 99, 6.19) # 0.3607524
qlambdap(0.55, 99, 6.19) # 6.31081

}
