% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CohensdpInitialize.R
\name{summarize}
\alias{summarize}
\title{summarize}
\usage{
summarize(x, ...)
}
\arguments{
\item{x}{an object to summarize}

\item{...}{ignored}
}
\description{
Summarize provides a human-readable output of a dpObject. it is 
synonyme of summary (but as actions are verbs, I used a verb).
}
