% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CohensdpInitialize.R, R/methods.R
\name{explain}
\alias{explain}
\alias{explain.CohensdpObject}
\title{explain}
\usage{
explain(x, ...)

\method{explain}{CohensdpObject}(x, ...)
}
\arguments{
\item{x}{an object to explain}

\item{...}{ignored}
}
\value{
a human-readable output with details of computations.
}
\description{
\code{explain()} provides a human-readable, exhaustive, description of
the results. It also provides references to the key results.
}
