% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CohensdpInitialize.R, R/methods.R
\name{summarize}
\alias{summarize}
\alias{summarize.CohensdpObject}
\title{summarize}
\usage{
summarize(x, ...)

\method{summarize}{CohensdpObject}(x, ...)
}
\arguments{
\item{x}{an object to summarize}

\item{...}{ignored}
}
\value{
a human-readable output as per articles.
}
\description{
\code{summarize()} provides a human-readable output of a dpObject. it is
synonym of \code{summary()} (but as actions are verbs, I used a verb).
}
